[feat/#12] 카카오 api 코드 리팩토링

+++ src/main/java/org/platemate/dto/request/PostUserDataRequest.java
@@ -2,7 +2,7 @@ package org.platemate.dto.request;
 
 public record PostUserDataRequest(
         String nickname,
-        Long latitude,
-        Long longtitude
+        Float latitude,
+        Float longtitude
 ) {
 }
